Use pattern matching

+++ src/Microsoft.CodeAnalysis.NetAnalyzers/src/Microsoft.CodeAnalysis.NetAnalyzers/Microsoft.CodeQuality.Analyzers/Maintainability/CodeMetricsAnalyzer.cs
@@ -298,7 +298,7 @@ namespace Microsoft.CodeQuality.Analyzers.Maintainability.CodeMetrics
 
                         CA1505RuleId => 10,
 
-                        CA1506RuleId => symbolKind == SymbolKind.NamedType ? 95 : (uint)40,
+                        CA1506RuleId => symbolKind == SymbolKind.NamedType ? 95 : 40,
 
                         _ => throw new NotImplementedException(),
                     };
